abstract edit and upvote to CommentInteraction
