+++ TourMate/TourMate/Frontend/Plan/Views/CommentViews/CommentView/CommentView.swift
@@ -18,11 +18,15 @@ struct CommentView: View {
         HStack(alignment: .bottom, spacing: 10.0) { // telegram style alignment
             CommentIconView(imageUrl: user.imageUrl)
 
-            CommentTextView(viewModel: viewModel, user: user, comment: comment)
-                .frame(maxWidth: .infinity, alignment: .leading)
-                .padding()
-                .background(.white)
-                .cornerRadius(20)
+            VStack(alignment: .leading, spacing: 10.0) {
+                CommentTextView(user: user, comment: comment)
+
+                CommentInteractionView(viewModel: viewModel, comment: comment)
+            }
+            .frame(maxWidth: .infinity, alignment: .leading)
+            .padding()
+            .background(.white)
+            .cornerRadius(20)
         }
     }
 }
